+++ apps/sr/bsafs/changelog.md
@@ -1,5 +1,13 @@
 # Version history of bsafs
 
+## Version 0.1.1 (2024-01-20)
+
+* The block size of a file is now reported correctly instead of always being
+  zero.
+* Furthermore, the time for last access, last modification and last status
+  change of files and directories in the archive are now set to the respective
+  value of the archive file instead of just reporting the current time.
+
 ## Version 0.1.0 (2024-01-20)
 
 This is the initial version of `bsafs`.